Added broadcast switch
Now users can pass {broadcast: false} with a request to persist it to the DB but avoid notifying other clients. This is useful if the other clients already know about the change because it's derived from another one but we want to have it in the DB for future bootstrapping.